move Tracker class to separate file
