+++ src/overlay/ItemFetcher.cpp
@@ -9,6 +9,7 @@
 #include "medida/metrics_registry.h"
 #include "herder/TxSetFrame.h"
 #include "overlay/StellarXDR.h"
+#include "overlay/Tracker.h"
 #include "crypto/Hex.h"
 #include "crypto/SHA.h"
 #include "herder/Herder.h"
@@ -16,9 +17,6 @@
 
 namespace stellar
 {
-
-static std::chrono::milliseconds const MS_TO_WAIT_FOR_FETCH_REPLY{1500};
-static int const MAX_REBUILD_FETCH_LIST = 1000;
 
 ItemFetcher::ItemFetcher(Application& app, AskPeer askPeer)
     : mApp(app)
@@ -57,7 +55,7 @@ ItemFetcher::isFetching(Hash itemHash) const
         return false;
     }
 
-    return iter->second->hasWaitingEnvelopes();
+    return !iter->second->empty();
 }
 
 void
@@ -109,173 +107,18 @@ ItemFetcher::recv(Hash itemHash)
     {
         // this code can safely be called even if recvSCPEnvelope ends up
         // calling recv on the same itemHash
-        auto& waiting = iter->second->mWaitingEnvelopes;
+        auto& tracker = iter->second;
 
         CLOG(TRACE, "Overlay") << "Recv " << hexAbbrev(itemHash) << " : "
-                               << waiting.size();
+                               << tracker->size();
 
-        while (!waiting.empty())
+        while (!tracker->empty())
         {
-            SCPEnvelope env = waiting.back().second;
-            waiting.pop_back();
-            mApp.getHerder().recvSCPEnvelope(env);
+            mApp.getHerder().recvSCPEnvelope(tracker->pop());
         }
         // stop the timer, stop requesting the item as we have it
-        iter->second->mTimer.cancel();
+        iter->second->cancel();
     }
-}
-
-Tracker::Tracker(Application& app, Hash const& hash, AskPeer &askPeer)
-    : mAskPeer(askPeer)
-    , mApp(app)
-    , mNumListRebuild(0)
-    , mTimer(app)
-    , mItemHash(hash)
-    , mTryNextPeerReset(app.getMetrics().NewMeter(
-          {"overlay", "item-fetcher", "reset-fetcher"}, "item-fetcher"))
-    , mTryNextPeer(app.getMetrics().NewMeter(
-          {"overlay", "item-fetcher", "next-peer"}, "item-fetcher"))
-{
-    assert(mAskPeer);
-}
-
-Tracker::~Tracker()
-{
-    mTimer.cancel();
-}
-
-// returns false if no one cares about this guy anymore
-bool
-Tracker::clearEnvelopesBelow(uint64 slotIndex)
-{
-    for (auto iter = mWaitingEnvelopes.begin();
-         iter != mWaitingEnvelopes.end();)
-    {
-        if (iter->second.statement.slotIndex < slotIndex)
-        {
-            iter = mWaitingEnvelopes.erase(iter);
-        }
-        else
-        {
-            iter++;
-        }
-    }
-    if (!mWaitingEnvelopes.empty())
-    {
-        return true;
-    }
-
-    mTimer.cancel();
-    mLastAskedPeer = nullptr;
-
-    return false;
-}
-
-void
-Tracker::doesntHave(Peer::pointer peer)
-{
-    if (mLastAskedPeer == peer)
-    {
-        CLOG(TRACE, "Overlay") << "Does not have " << hexAbbrev(mItemHash);
-        tryNextPeer();
-    }
-}
-
-void
-Tracker::tryNextPeer()
-{
-    // will be called by some timer or when we get a
-    // response saying they don't have it
-    Peer::pointer peer;
-
-    CLOG(TRACE, "Overlay") << "tryNextPeer " << hexAbbrev(mItemHash) << " last: "
-                           << (mLastAskedPeer ? mLastAskedPeer->toString()
-                                              : "<none>");
-
-    // if we don't have a list of peers to ask and we're not
-    // currently asking peers, build a new list
-    if (mPeersToAsk.empty() && !mLastAskedPeer)
-    {
-        std::set<std::shared_ptr<Peer>> peersWithEnvelope;
-        for (auto const& e : mWaitingEnvelopes)
-        {
-            auto const& s = mApp.getOverlayManager().getPeersKnows(e.first);
-            peersWithEnvelope.insert(s.begin(), s.end());
-        }
-
-        // move the peers that have the envelope to the back,
-        // to be processed first
-        for (auto const& p : mApp.getOverlayManager().getRandomPeers())
-        {
-            if (peersWithEnvelope.find(p) != peersWithEnvelope.end())
-            {
-                mPeersToAsk.emplace_back(p);
-            }
-            else
-            {
-                mPeersToAsk.emplace_front(p);
-            }
-        }
-
-        mNumListRebuild++;
-
-        CLOG(TRACE, "Overlay") << "tryNextPeer " << hexAbbrev(mItemHash)
-                               << " attempt " << mNumListRebuild
-                               << " reset to #" << mPeersToAsk.size();
-        mTryNextPeerReset.Mark();
-    }
-
-    while (!peer && !mPeersToAsk.empty())
-    {
-        peer = mPeersToAsk.back();
-        if (!peer->isAuthenticated())
-        {
-            peer.reset();
-        }
-        mPeersToAsk.pop_back();
-    }
-
-    std::chrono::milliseconds nextTry;
-    if (!peer)
-    { // we have asked all our peers
-        // clear mLastAskedPeer so that we rebuild a new list
-        mLastAskedPeer.reset();
-        if (mNumListRebuild > MAX_REBUILD_FETCH_LIST)
-        {
-            nextTry = MS_TO_WAIT_FOR_FETCH_REPLY * MAX_REBUILD_FETCH_LIST;
-        }
-        else
-        {
-            nextTry = MS_TO_WAIT_FOR_FETCH_REPLY * mNumListRebuild;
-        }
-    }
-    else
-    {
-        mLastAskedPeer = peer;
-        CLOG(TRACE, "Overlay") << "Asking for " << hexAbbrev(mItemHash) << " to "
-                               << peer->toString();
-        mTryNextPeer.Mark();
-        mAskPeer(peer, mItemHash);
-        nextTry = MS_TO_WAIT_FOR_FETCH_REPLY;
-    }
-
-    mTimer.expires_from_now(nextTry);
-    mTimer.async_wait(
-        [this]()
-        {
-            this->tryNextPeer();
-        },
-        VirtualTimer::onFailureNoop);
-}
-
-void
-Tracker::listen(const SCPEnvelope& env)
-{
-    StellarMessage m;
-    m.type(SCP_MESSAGE);
-    m.envelope() = env;
-    mWaitingEnvelopes.push_back(
-        std::make_pair(sha256(xdr::xdr_to_opaque(m)), env));
 }
 
 }
